+++ pyhdc/encodings/matrix.py
@@ -1,4 +1,4 @@
-﻿#!/usr/bin/env python
+#!/usr/bin/env python
 """
 Matrix-Based Encodings for HDC